@@ -4,7 +4,9 @@
       <div class="row">
         <div class="col-md-8 col-md-offset-2 footer">
           <div class="copyright">
-            <small>Copyright &copy; Yin&Yang Calendar All rights reserved.</small>
+            <small
+              >Copyright &copy; Yin&Yang Calendar All rights reserved.</small
+            >
           </div>
         </div>
       </div>
@@ -12,13 +14,7 @@
   </footer>
 </template>
 
-<style lang="scss">
-@import '~assets/main.scss';
-</style>
-
 <style lang="scss" scoped>
-@import '~assets/main.scss';
-
 .footer {
   margin-top: 25px;
   margin-bottom: 25px;
