@@ -4,45 +4,53 @@
       <div class="container text-center">
         <div class="row">
           <div class="col-md-8 col-md-offset-2 main">
-            <div class="moon">
-              <img 
-                src="~assets/moon@2x.png" 
-                alt >
-            </div>
+            <!-- <div class="moon">
+              <img src="~assets/moon.svg" alt />
+            </div> -->
             <div class="title">
-              <img 
-                src="~assets/title@2x.png" 
-                alt="Yin&Yang Calendar 2018" >
+              <img src="~assets/title.svg" alt="Yin&Yang Calendar 2018" />
             </div>
             <div class="text">
-              <img 
-                src="~assets/text@2x.png" 
-                alt="和暦は月と太陽のリズムを組み合わせた、「太陽太陰暦」。" >
+              <img
+                src="~assets/text.svg"
+                alt="和暦は月と太陽のリズムを組み合わせた、「太陽太陰暦」。"
+              />
             </div>
           </div>
           <div class="col-md-8 col-md-offset-2 banner">
             <ul class="list-inline">
               <li class="center-block">
                 <a
-                  href="https://www.amazon.co.jp/o/ASIN/B07YTF5W5F/chiakira-22/"
+                  href="https://www.amazon.co.jp/o/ASIN/B08JH7XT5H/chiakira-22/"
                   target="_blank"
-                  class="btn btn-default btn-lg"
+                  class="btn btn-default btn-lg color"
                 >
-                  <i class="fa fa-amazon fa-fw" />
-                  <span class="network-name">Amazon.co.jpで購入する</span>
+                  <i class="fab fa-amazon fa-fw"></i>
+                  <span class="network-name">Amazonで詳しくみる</span>
                 </a>
               </li>
-              <div class>
-                <nuxt-link to="/pricing/">10部以上ご購入の方へ</nuxt-link>
+              <div>
+                <nuxt-link to="/pricing/">卸価格について</nuxt-link>
               </div>
-              <li>
+              <li class="center-block">
                 <a
-                  href="https://ameblo.jp/tukiyomichan/"
+                  href="https://tukiyomi.in/"
                   target="_blank"
                   class="btn btn-default btn-lg"
                 >
-                  <i class="fa fa-pencil-square-o fa-fw" />
-                  <span class="network-name">Blog</span>
+                  <span class="network-name">
+                    <small>Salon</small>つきよみ白金台
+                  </span>
+                </a>
+              </li>
+              <li>
+                <a
+                  href="https://www.instagram.com/tukiyomichan/"
+                  target="_blank"
+                  class="btn btn-default btn-lg"
+                >
+                  <i class="fab fa-instagram fa-fw"></i>
+                  <span class="network-name">Instagram</span>
                 </a>
               </li>
               <li>
@@ -51,7 +59,7 @@
                   target="_blank"
                   class="btn btn-default btn-lg"
                 >
-                  <i class="fa fa-facebook-official fa-fw" />
+                  <i class="fab fa-facebook-square fa-fw"></i>
                   <span class="network-name">Facebook</span>
                 </a>
               </li>
@@ -61,18 +69,18 @@
                   target="_blank"
                   class="btn btn-default btn-lg"
                 >
-                  <i class="fa fa-twitter fa-fw" />
+                  <i class="fab fa-twitter fa-fw"></i>
                   <span class="network-name">Twitter</span>
                 </a>
               </li>
               <li>
-                <a 
-                  href="http://tukiyomi.in/" 
-                  target="_blank" 
-                  class="btn btn-default btn-lg">
-                  <span class="network-name">
-                    <small>Salon</small>つきよみ白金台
-                  </span>
+                <a
+                  href="https://ameblo.jp/tukiyomichan/"
+                  target="_blank"
+                  class="btn btn-default btn-lg"
+                >
+                  <i class="fas fa-edit fa-fw"></i>
+                  <span class="network-name">Blog</span>
                 </a>
               </li>
             </ul>
@@ -91,11 +99,7 @@ import FixHeight from '~/components/FixHeight.vue'
 export default {
   components: {
     GlobalFooter,
-    FixHeight
-  }
+    FixHeight,
+  },
 }
 </script>
-
-<style lang="scss">
-@import '~assets/main.scss';
-</style>
